requirements.txt
numpy
pillow
matplotlib
scikit-learn
scipy
torch
torchvision